+++ Changelog.rst
@@ -17,6 +17,7 @@ Resolved issues
 
 * RSA key size returned correctly in RsaKey.__repr__ method.
 * CTR mode does not modify anymore 'counter' parameter passed to 'new' method.
+* CTR raises OverflowError instead of ValueError in case of counter wrapping around.
 
 Breaks in compatibility
 -----------------------